requirements.txt
pika==1.3.2
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0

torch==2.0.1
transformers==4.32.1
pillow==10.0.0

python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.0

fastapi==0.115.12
pip==24.3.1
pipreqs==0.5.0
tinycss2==1.4.0
uvicorn==0.34.0
sqlalchemy==2.0.29
asyncpg==0.29.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
pydantic==2.7.1
bcrypt==4.1.2
